Add functionality to retrieve 'date' from Microsoft API
This commit adds the ability to retrieve the 'date' parameter from the Microsoft API using the provided 'name'. If a valid response is received, the retrieved value is assigned to 'date'. Otherwise, an error message is displayed. This change enhances the Azure Function by allowing it to dynamically fetch data based on the provided input.

Also, update README.md with instructions for looking up a CVE without knowing the date. The new request body format and example are added to facilitate this functionality.

Sources:
- [Microsoft Security Updates](https://msrc.microsoft.com/update-guide/)
- [Microsoft Security Developers Guide](https://api.msrc.microsoft.com/cvrf/v2.0/swagger/index)